+++ packages/orq-rc/src/models/operations/retrievedataset.ts
@@ -126,7 +126,7 @@ export const RetrieveDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-04-12T10:25:48.200Z",
+    "2026-04-13T03:39:40.773Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {