+++ packages/orq-rc/RELEASES.md
@@ -8219,3 +8219,13 @@ Based on:
 - [typescript v4.8.0-rc.6] packages/orq-rc
 ### Releases
 - [NPM v4.8.0-rc.6] https://www.npmjs.com/package/@orq-ai/node/v/4.8.0-rc.6 - packages/orq-rc
+
+## 2026-04-13 03:42:49
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.3 (2.879.11) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.8.0-rc.7] packages/orq-rc
+### Releases
+- [NPM v4.8.0-rc.7] https://www.npmjs.com/package/@orq-ai/node/v/4.8.0-rc.7 - packages/orq-rc